+++ CHANGELOG.md
@@ -4,7 +4,7 @@ https://github.com/jcorporation/myMPD/
 
 ***
 
-## myMPD v23.0.0 (not yet released)
+## myMPD v23.0.0 (2025-11-09)
 
 This versions enhances the jukebox implementation and the album handling.